+++ NEWS.md
@@ -5,7 +5,8 @@ Bug fixes and minor improvements
 --------------------------------
 
 * Skip a test on CRAN which was failing on `r-oldrel-windows-ix86+x86_64` (#52)
-
+* New `as_datetimeoffset.integer()` and `as_datetimeoffset.numeric()`
+  coerce numeric values as the new `datetimeoffset()` objects `year` field (#53).
 
 datetimeoffset 0.1.2
 ====================